backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database & ORM
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Data Validation & Serialization
pydantic==2.5.0
pydantic-settings==2.1.0

# CORS & Security
python-multipart==0.0.6

# Development
python-dotenv==1.0.0 